python-sanic-asyncpg/requirements.txt
sanic
uvicorn[standard]
asyncpg
python-dotenv
black
